Increased test coverage and fixed a potential rounding problem
The previous implementation always returned zero for any era lasting less than 2.5 minutes.

After this change it works for smaller eras as well.

There may still be minor rounding issues during the division, but it's fine as this payout mechanism is temporary and works for era times that are multiple of 6 seconds (current block time), which is a divisor of 150 seconds (Zend block time).

+++ runtime/src/lib.rs
@@ -335,9 +335,9 @@ impl pallet_staking::EraPayout<Balance> for ZendPayout {
         _total_issuance: Balance,
         era_duration_millis: u64,
     ) -> (Balance, Balance) {
-        const HORIZEN_POW_MILLISECS_PER_BLOCK: u64 = 150 * 1000; // 2.5 minutes
+        const HORIZEN_POW_MILLISECS_PER_BLOCK: u128 = 150 * 1000; // 2.5 minutes
         let era_reward: u128 =
-            625 * CENTS * u128::from(era_duration_millis / HORIZEN_POW_MILLISECS_PER_BLOCK); // 6.25nZEN per block
+            625 * CENTS * u128::from(era_duration_millis) / HORIZEN_POW_MILLISECS_PER_BLOCK; // 6.25nZEN per Zend block
 
         // Assign 100% of the reward to validators.
         // In the future we may want to split the reward between validators and the treasury.
